# Optimize scalar lists