--- a/src/GraphQL.Tests/Serialization/ExecutionResultTests.cs
+++ b/src/GraphQL.Tests/Serialization/ExecutionResultTests.cs
@@ -1,4 +1,6 @@
+using System.Numerics;
 using GraphQL.Tests.StarWars;
+using GraphQL.Types;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace GraphQL.Tests.Serialization;
@@ -171,5 +173,94 @@ public class ExecutionResultTests
         await serializer.WriteAsync(stream, result);
         string asyncResult = System.Text.Encoding.UTF8.GetString(stream.ToArray());
         syncResult.ShouldBe(asyncResult);
+    }
+
+    [Theory]
+    [ClassData(typeof(GraphQLSerializersTestData))]
+    public async Task IntegralListTypesSerializeCorrectly(IGraphQLTextSerializer serializer)
+    {
+        var schema = new Schema
+        {
+            Query = new AutoRegisteringObjectGraphType<ArraySampleQuery>(),
+        };
+        schema.Initialize();
+        var result = await new DocumentExecuter().ExecuteAsync(new ExecutionOptions
+        {
+            Schema = schema,
+            Query = "{ bools bools2 sBytes sBytes2 bytes bytes2 bytes3 shorts shorts2 uShorts uShorts2 ints ints2 ints3 uInts uInts2 longs longs2 uLongs uLongs2 singles singles2 doubles doubles2 decimals decimals2 bigIntegers bigIntegers2 strings strings2 }"
+        });
+        string actual = serializer.Serialize(result);
+        actual.ShouldBeCrossPlatJson("""
+            {
+              "data": {
+                "bools": [true, false],
+                "bools2": [true, null],
+                "sBytes": [-1, 0, 1],
+                "sBytes2": [-1, null, 1],
+                "bytes": [0, 1, 2],
+                "bytes2": [null, 1, 2],
+                "bytes3": [0, 1, 2],
+                "shorts": [-1, 0, 1],
+                "shorts2": [-1, null, 1],
+                "uShorts": [0, 1, 2],
+                "uShorts2": [null, 1, 2],
+                "ints": [-1, 0, 1],
+                "ints2": [-1, null, 1],
+                "ints3": [-1, 0, 1],
+                "uInts": [0, 1, 2],
+                "uInts2": [null, 1, 2],
+                "longs": [-1, 0, 1],
+                "longs2": [-1, null, 1],
+                "uLongs": [0, 1, 2],
+                "uLongs2": [null, 1, 2],
+                "singles": [-1.5, 0, 1.5],
+                "singles2": [-1.5, null, 1.5],
+                "doubles": [-1.5, 0, 1.5],
+                "doubles2": [-1.5, null, 1.5],
+                "decimals": [-1.5, 0, 1.5],
+                "decimals2": [-1.5, null, 1.5],
+                "bigIntegers": [-1, 0, 1],
+                "bigIntegers2": [-1, null, 1],
+                "strings": ["abc", "def", "ghi"],
+                "strings2": ["abc", null, "ghi"]
+              }
+            }
+            """);
+    }
+
+    private class ArraySampleQuery()
+    {
+        public static bool[] Bools => [true, false];
+        public static bool?[] Bools2 => [true, null];
+
+        public static sbyte[] SBytes => [-1, 0, 1];
+        public static sbyte?[] SBytes2 => [-1, null, 1];
+        public static byte[] Bytes => [0, 1, 2];
+        public static byte?[] Bytes2 => [null, 1, 2];
+        public static List<byte> Bytes3 => [0, 1, 2];
+        public static short[] Shorts => [-1, 0, 1];
+        public static short?[] Shorts2 => [-1, null, 1];
+        public static ushort[] UShorts => [0, 1, 2];
+        public static ushort?[] UShorts2 => [null, 1, 2];
+        public static int[] Ints => [-1, 0, 1];
+        public static int?[] Ints2 => [-1, null, 1];
+        [OutputType(typeof(NonNullGraphType<ListGraphType<NonNullGraphType<IntGraphType>>>))]
+        public static long[] Ints3 => [-1, 0, 1];
+        public static uint[] UInts => [0, 1, 2];
+        public static uint?[] UInts2 => [null, 1, 2];
+        public static long[] Longs => [-1, 0, 1];
+        public static long?[] Longs2 => [-1, null, 1];
+        public static ulong[] ULongs => [0, 1, 2];
+        public static ulong?[] ULongs2 => [null, 1, 2];
+        public static float[] Singles => [-1.5f, 0, 1.5f];
+        public static float?[] Singles2 => [-1.5f, null, 1.5f];
+        public static double[] Doubles => [-1.5, 0, 1.5];
+        public static double?[] Doubles2 => [-1.5, null, 1.5];
+        public static decimal[] Decimals => [-1.5m, 0, 1.5m];
+        public static decimal?[] Decimals2 => [-1.5m, null, 1.5m];
+        public static BigInteger[] BigIntegers => [-1, 0, 1];
+        public static BigInteger?[] BigIntegers2 => [-1, null, 1];
+        public static string[] Strings => ["abc", "def", "ghi"];
+        public static string?[] Strings2 => ["abc", null, "ghi"];
     }
 }