--- a/src/GraphQL/Types/Scalars/IdGraphType.cs
+++ b/src/GraphQL/Types/Scalars/IdGraphType.cs
@@ -1,3 +1,5 @@
+using System.Collections;
+using System.Globalization;
 using System.Numerics;
 using GraphQLParser.AST;
 
@@ -62,5 +64,39 @@ public class IdGraphType : ScalarGraphType
     };
 
     /// <inheritdoc/>
-    public override object? Serialize(object? value) => ParseValue(value)?.ToString();
+    public override object? Serialize(object? value) => value switch
+    {
+        string _ => value,
+        int i => i.ToString(CultureInfo.InvariantCulture),
+        long l => l.ToString(CultureInfo.InvariantCulture),
+        Guid g => g.ToString("D"),
+        null => null,
+        byte b => b.ToString(CultureInfo.InvariantCulture),
+        sbyte sb => sb.ToString(CultureInfo.InvariantCulture),
+        short s => s.ToString(CultureInfo.InvariantCulture),
+        ushort us => us.ToString(CultureInfo.InvariantCulture),
+        uint ui => ui.ToString(CultureInfo.InvariantCulture),
+        ulong ul => ul.ToString(CultureInfo.InvariantCulture),
+        BigInteger bi => bi.ToString(CultureInfo.InvariantCulture),
+        _ => ThrowSerializationError(value)
+    };
+
+    /// <inheritdoc/>
+    public override bool CanSerializeList(IEnumerable list, bool allowNulls) =>
+        CanSerializeList<int>(list, allowNulls) ||
+        CanSerializeList<long>(list, allowNulls) ||
+        CanSerializeList<Guid>(list, allowNulls) ||
+        list is IEnumerable<string?> values && values.FastAll(static s => s != null);
+
+    /// <inheritdoc/>
+    public override IEnumerable SerializeList(IEnumerable list) => list switch
+    {
+        IEnumerable<int> values => values.Select(value => value.ToString(CultureInfo.InvariantCulture)),
+        IEnumerable<int?> values => values.Select(value => value?.ToString(CultureInfo.InvariantCulture)),
+        IEnumerable<long> values => values.Select(value => value.ToString(CultureInfo.InvariantCulture)),
+        IEnumerable<long?> values => values.Select(value => value?.ToString(CultureInfo.InvariantCulture)),
+        IEnumerable<Guid> values => values.Select(value => value.ToString("D")),
+        IEnumerable<Guid?> values => values.Select(value => value?.ToString("D")),
+        _ => throw new NotSupportedException(),
+    };
 }