… src/GraphQL.ApiTests/net50/GraphQL.approved.txt
… src/GraphQL.ApiTests/net60/GraphQL.approved.txt
… src/GraphQL.ApiTests/netstandard20+netstandard21/GraphQL.approved.txt